+++ workos/typing/webhooks.py
@@ -1,7 +1,7 @@
 from typing import Any, Dict, Union
 from typing_extensions import Annotated
 from pydantic import Field, TypeAdapter
-from workos.resources.webhooks import Webhook
+from workos.types.webhooks.webhook import Webhook
 from workos.types.workos_model import WorkOSModel
