+++ workos/webhooks.py
@@ -3,7 +3,7 @@ import hmac
 import time
 import hashlib
 from typing import Optional, Protocol
-from workos.resources.webhooks import Webhook
+from workos.types.webhooks.webhook import Webhook
 from workos.types.webhooks.webhook_payload import WebhookPayload
 from workos.typing.webhooks import WebhookTypeAdapter
 from workos.utils.validation import Module, validate_settings
